committing sum task

+++ sum/sum.h
@@ -3,6 +3,6 @@
 
 #include <cstdint>
 
-int64_t Sum(int32_t x, int32_t y);
+int64_t Sum(int64_t x, int64_t y);
 
-#endif  // SUM_SUM_H
+#endif
